Fix based on review feedback
