+++ src/parse/mod.rs
@@ -1,5 +1,6 @@
 //! Parsers for the various kinds of game script.
 
+pub mod cob;
 #[cfg(any(feature = "ck3", feature = "imperator"))]
 pub mod csv;
 #[cfg(feature = "vic3")]
